third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-animation-corner-cases.tentative.html


<!DOCTYPE html>
<meta charset="utf-8">
<link rel=author href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/popup-utils.js"></script>

<body>
<style>
.animation { opacity: 0; }
.animation:top-layer { opacity: 1; }
.animation:not(:top-layer) { animation: fade-out 1000s; }
@keyframes fade-out {
  from { opacity: 1; }
  to { opacity: 0; }
}

.animation>div>div { left: 0; }
.animation:not(:top-layer)>div>div { animation: rotate 1000s; color:red;}
@keyframes rotate {
  from { transform: rotate(0); }
  to { transform: rotate(360deg); }
}

[popup] { top: 200px; }
[popup]::backdrop { background-color: rgba(255,0,0,0.2); }
</style>

<script>
function createPopUp(t,type) {
  const popUp = document.createElement('div');
  popUp.popUp = 'auto';
  popUp.classList = type;
  const div = document.createElement('div');
  const descendent = div.appendChild(document.createElement('div'));
  descendent.appendChild(document.createTextNode("Descendent element"));
  popUp.append("This is a pop up",div);
  document.body.appendChild(popUp);
  t.add_cleanup(() => popUp.remove());
  return {popUp, descendent};
}
promise_test(async (t) => {
  const {popUp, descendent} = createPopUp(t,'animation');
  assert_false(isElementVisible(popUp));
  assert_equals(descendent.parentElement.parentElement,popUp);
  popUp.showPopUp();
  assert_true(popUp.matches(':top-layer'));
  assert_true(isElementVisible(popUp));
  assert_equals(popUp.getAnimations({subtree: true}).length,0);
  popUp.hidePopUp();
  const animations = popUp.getAnimations({subtree: true});
  assert_equals(animations.length,2,'There should be two animations running');
  assert_false(popUp.matches(':top-layer'),'popUp should not match :top-layer as soon as hidden');
  assert_true(isElementVisible(popUp),'but animations should keep the popUp visible');
  assert_true(isElementVisible(descendent),'The descendent should also be visible');
  await waitForRender();
  await waitForRender();
  assert_equals(popUp.getAnimations({subtree: true}).length,2,'The animations should still be running');
  assert_true(isElementVisible(popUp),'PopUp should still be visible due to animation');
  animations.forEach(animation => animation.finish()); // Force the animations to finish
  await waitForRender(); // Wait one frame
  assert_false(popUp.matches(':top-layer'),'The pop up still shouldn\'t match :top-layer');
  assert_false(isElementVisible(popUp),'The pop up should now be invisible');
  assert_false(isElementVisible(descendent),'The descendent should also be invisible');
  assert_equals(popUp.getAnimations({subtree: true}).length,0);
},'Descendent animations should keep the pop up visible until the animation ends');

promise_test(async (t) => {
  const {popUp, descendent} = createPopUp(t,'');
  assert_equals(popUp.classList.length, 0);
  assert_false(isElementVisible(popUp));
  popUp.showPopUp();
  assert_true(popUp.matches(':top-layer'));
  assert_true(isElementVisible(popUp));
  assert_equals(popUp.getAnimations({subtree: true}).length,0);
  // Start an animation on the popUp and its descendent.
  popUp.animate([{opacity: 1},{opacity: 0}],{duration: 1000000,iterations: 1});
  descendent.animate([{transform: 'rotate(0)'},{transform: 'rotate(360deg)'}],{duration: 1000000,iterations: 1});
  assert_equals(popUp.getAnimations({subtree: true}).length,2);
  // Then hide the popUp.
  popUp.hidePopUp();
  assert_false(popUp.matches(':top-layer'),'pop up should not match :top-layer as soon as hidden');
  assert_equals(popUp.getAnimations({subtree: true}).length,2,'animations should still be running');
  await waitForRender();
  assert_equals(popUp.getAnimations({subtree: true}).length,2,'animations should still be running');
  assert_false(isElementVisible(popUp),'Pre-existing animations should not keep the pop up visible');
},'Pre-existing animations should *not* keep the pop up visible until the animation ends');
</script>
